change path && dataset

+++ models/PaddleGAN/CI/PaddleGAN_CI_develop.sh
@@ -52,7 +52,7 @@ else
 fi
 done
 
-find configs/ -name *.yaml -exec ls -l {} \; | awk '{print $NF;}'| grep -v 'wav2lip' | grep -v 'lapstyle_rev_second' | grep -v 'lapstyle_rev_first'| grep -v 'firstorder_vox_256'> models_list_all
+find configs/ -name *.yaml -exec ls -l {} \; | awk '{print $NF;}'| grep -v 'wav2lip' | grep -v 'lapstyle_rev_second' | grep -v 'lapstyle_rev_first'| grep -v 'firstorder_vox_256' | grep -v 'lapstyle_draft' > models_list_all
 git diff $(git log --pretty=oneline |grep "#"|head -1|awk '{print $1}') HEAD --diff-filter=AMR | grep diff|grep yaml|awk -F 'b/' '{print$2}'| tee -a  models_list
 shuf models_list_all >> models_list
 wc -l models_list
@@ -63,7 +63,9 @@ do
 echo $line
 filename=${line##*/}
 model=${filename%.*}
-rm -rf output
+if [ -d "output" ]; then
+   rm -rf output
+fi
 sed -i '1s/epochs/total_iters/' $line
 # animeganv2
 sed -i 's/pretrain_ckpt:/pretrain_ckpt: #/g' $line
@@ -71,9 +73,11 @@ export CUDA_VISIBLE_DEVICES=${cudaid2}
 case ${model} in
 lapstyle_rev_second)
 python tools/main.py --config-file $line -o total_iters=20 snapshot_config.interval=10 log_config.interval=1 output_dir=output dataset.train.batch_size=2 > $log_path/train/${model}.log 2>&1
+ls
 params_dir=$(ls output)
+echo "params_dir"
+echo $params_dir
 if [ -f "output/$params_dir/iter_20_checkpoint.pdparams" ];then
-   cat $log_path/train/${model}.log
    echo -e "\033[33m train of $model  successfully!\033[0m"| tee -a $log_path/result.log
 else
    cat $log_path/train/${model}.log
@@ -93,7 +97,6 @@ esac
 
 # evaluate 
 export CUDA_VISIBLE_DEVICES=${cudaid1}
-echo $params_dir
 ls output/$params_dir/
 case ${model} in
 stylegan_v2_256_ffhq)
